requirements.txt
anyio==3.6.2
certifi==2023.5.7
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
psycopg2-binary==2.9.6
python-dotenv==1.0.0
python-telegram-bot==20.3
sniffio==1.3.0
SQLAlchemy==2.0.13
SQLAlchemy-Utils==0.41.1
typing_extensions==4.5.0
